+++ app/src/main/java/org/simple/clinic/settings/SettingsEffectHandler.kt
@@ -8,6 +8,8 @@ import dagger.assisted.AssistedInject
 import io.reactivex.ObservableTransformer
 import org.simple.clinic.appupdate.AppUpdateState
 import org.simple.clinic.appupdate.CheckAppUpdateAvailability
+import org.simple.clinic.patient.PatientRepository
+import org.simple.clinic.patient.medicalRecords.PushMedicalRecordsOnline
 import org.simple.clinic.storage.DatabaseEncryptor
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.filterAndUnwrapJust
@@ -16,10 +18,12 @@ import org.simple.clinic.util.scheduler.SchedulersProvider
 class SettingsEffectHandler @AssistedInject constructor(
     private val userSession: UserSession,
     private val settingsRepository: SettingsRepository,
+    private val patientRepository: PatientRepository,
     private val schedulersProvider: SchedulersProvider,
     private val appVersionFetcher: AppVersionFetcher,
     private val appUpdateAvailability: CheckAppUpdateAvailability,
     private val databaseEncryptor: DatabaseEncryptor,
+    private val pushMedicalRecordsOnline: PushMedicalRecordsOnline,
     @Assisted private val viewEffectsConsumer: Consumer<SettingsViewEffect>
 ) {
 
@@ -39,6 +43,8 @@ class SettingsEffectHandler @AssistedInject constructor(
       .addConsumer(SettingsViewEffect::class.java, viewEffectsConsumer::accept)
       .addTransformer(LogoutUser::class.java, logoutUser())
       .addTransformer(LoadDatabaseEncryptionStatus::class.java, loadDatabaseEncryptionStatus())
+      .addTransformer(FetchCompleteMedicalRecords::class.java, fetchCompleteMedicalRecords())
+      .addTransformer(PushCompleteMedicalRecordsOnline::class.java, pushCompleteMedicalRecordsOnline())
       .build()
 
   private fun loadDatabaseEncryptionStatus(): ObservableTransformer<LoadDatabaseEncryptionStatus, SettingsEvent> {
@@ -94,6 +100,26 @@ class SettingsEffectHandler @AssistedInject constructor(
           }
           .map {
             AppUpdateAvailabilityChecked(it is AppUpdateState.ShowAppUpdate)
+          }
+    }
+  }
+
+  private fun fetchCompleteMedicalRecords(): ObservableTransformer<FetchCompleteMedicalRecords, SettingsEvent> {
+    return ObservableTransformer { effects ->
+      effects
+          .observeOn(schedulersProvider.io())
+          .map { patientRepository.fetchCompleteMedicalRecord() }
+          .map(::MedicalRecordsFetched)
+    }
+  }
+
+  private fun pushCompleteMedicalRecordsOnline(): ObservableTransformer<PushCompleteMedicalRecordsOnline, SettingsEvent> {
+    return ObservableTransformer { effects ->
+      effects
+          .observeOn(schedulersProvider.io())
+          .map {
+            val results = pushMedicalRecordsOnline.pushAllMedicalRecordsOnServer(it.medicalRecords)
+            PushMedicalRecordsOnlineCompleted(results)
           }
     }
   }